requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
python-dateutil==2.9.0.post0
tqdm==4.66.4
requests==2.32.3
regex==2024.5.15

# Google APIs (YouTube + CSE)
google-api-python-client==2.131.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# NLP sentiment (Hugging Face)
transformers==4.41.2
torch==2.3.1
sentencepiece==0.2.0

# App
streamlit==1.36.0
altair==5.3.0

# Tests
pytest==8.2.0
